refactoring config reading to allow for multiple configurations
